added monitor config to hyprland
I'm not sure if the hyprland config works, because the mirroring
of the monitor over the HDMI port does not work. It seems like
that `wayland.windowManager.hyprland.settings` simply not works.
I hope to check this out in the future and to get a clean hyprland
config in nix syntax.