fixed hyprland config in nix syntax
I had a bad bug inside my config, but now I found out how to fix
it. The bug was that I forgot to enable the option
wayland.wandowManager.hyprland .
Because of this, the settings defined inside were not be used,
that is why I kept the .conf files inside my config. But now it
works seemlessly and I am very happy.