removed pam for greetd
setting the pam for greetd is not a good idea, it leads to various
unexpected error. A big one is that hyprland cannot be started, because
pam sets the XDG_RUNTIME_DIR, which is neccessary for hyprland to work.
Otherways it just crashes. Therefore there is no 2fa for login by now,
but this will be added soon when I switch to sddm or similar tools.